Refactored QueryBuilder for fluent usage
