+++ commons-rest/src/main/java/io/tackle/commons/resources/query/QueryBuild.java
@@ -1,0 +1,7 @@
+package io.tackle.commons.resources.query;
+
+import io.quarkus.hibernate.orm.panache.PanacheEntity;
+
+public interface QueryBuild<ENTITY extends PanacheEntity> {
+    Query build();
+}
