+++ commons-rest/src/main/java/io/tackle/commons/resources/query/QueryUriInfo.java
@@ -1,0 +1,9 @@
+package io.tackle.commons.resources.query;
+
+import io.quarkus.hibernate.orm.panache.PanacheEntity;
+
+import javax.ws.rs.core.UriInfo;
+
+public interface QueryUriInfo<ENTITY extends PanacheEntity> {
+    QueryBuild<ENTITY> andUriInfo(UriInfo uriInfo);
+}
